remove unused files andupdate references
