+++ frontend/src/components/search/SearchPaginationFetch.tsx
@@ -1,7 +1,7 @@
 "use server";
 
 import fetchers from "src/app/api/Fetchers";
-import { QueryParamData } from "src/services/search/searchfetcher/SearchFetcher";
+import { QueryParamData } from "src/types/search/searchRequestTypes";
 
 import SearchPagination from "src/components/search/SearchPagination";
